+++ src/main/java/io/github/astrapi69/crypto/algorithm/KeyPairWithModeAndPaddingAlgorithm.java
@@ -36,6 +36,13 @@ public enum KeyPairWithModeAndPaddingAlgorithm implements Algorithm
 {
 
 	/**
+	 * The enum constant with the 'AES' algorithm with mode 'CBC' and padding 'PKCS5Padding' that
+	 * results in 'AES/CBC/PKCS5Padding'.
+	 */
+	AES_CBC_PKCS5Padding(AesAlgorithm.AES.name() + "/" + Mode.CBC.name() + "/"
+		+ PaddingScheme.PKCS5Padding.name()),
+
+	/**
 	 * The enum constant with the 'DESede' algorithm with mode 'CBC' and padding 'PKCS5Padding' that
 	 * results in 'DESede/CBC/PKCS5Padding'.
 	 */